Added health and damage system
Implemented interfaces for damageable and damaging game objects. Created health and damage scripts which implement these as well as a specific sword damage and box health script to deal with damaging collisions. In the sword damage script functionality was added to increase the damage amount up to a maximum based on the swing distance calculated in the momentum tracker script. This maximum distance is the maximum distance travelled by the hand in the 0.3 seconds prior to the moment it is calculated.