State Machine Update
Player state machine updated to be a separate singleton. IState interface created for states to inherit from, as well as a base player state which requires the player controller for the constructor. Separate states were created for walking, sprinting, crouching and dashing to control the enter executre and exit logic for each state as well as conditions for when the state can be entered or exited. Methods were created in the player controller to tell it what to change about the player behaviour and at which stage of execution to make the  changes. these were then implemented in the specific states where they were required.