Blocking enemy created
created a stationary enemy which blocks all incoming attacks from game objects using components which extend the IDamaging interface. This involved two steps, firstly moving the enemies sword to the same position in 3D space as the incoming game object. Secondly, rotating the sword to be perpendicular to the angle of attack of the incoming game object.