Fixed player health and damage
Added health script to the player as well as a collider and rigidbody to allow for collisions from the enemy sword. Also created a script which resizes and recenters the capsule collider to match the character controller collider. Added additional animation for horizontal slashing as well as logic in the enemy controller to randomise which counter attack to use.